# api/openapi.yaml
openapi: 3.1.0
info:
  title: HeatCheck Public API
  version: 1.0.0
  description: |
    Production API for HeatCheck challenges, proof clips, OpenAI-assisted
    verification, paid participation, voting, moderation, account lifecycle,
    and copyright workflows. All timestamps use RFC 3339 UTC values.
servers:
  - url: https://heatcheck.dogi.watch
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /v1/auth/verify-email:
    post:
      tags: [Authentication]
      operationId: verifyEmail
      summary: Verify an email address
      requestBody:
        $ref: "#/components/requestBodies/OpaqueToken"
      responses:
        "204":
          description: Email address verified.
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      summary: Request a password reset
      description: Always returns an accepted response to prevent account enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Password-reset request accepted.
        "429":
          $ref: "#/components/responses/Error"
  /v1/auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset an account password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                  writeOnly: true
                new_password:
                  type: string
                  minLength: 10
                  maxLength: 128
                  writeOnly: true
      responses:
        "204":
          description: Password changed and all sessions revoked.
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
...
  /v1/me/email-verification:
    post:
      tags: [Authentication]
      operationId: resendEmailVerification
      summary: Resend email verification
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Verification email queued.
        "204":
          description: Email was already verified.
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /v1/me/sessions:
    get:
      tags: [Authentication]
      operationId: listSessions
      summary: List active account sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active refresh-token families.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListEnvelope"
        "401":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Authentication]
      operationId: revokeAllSessions
      summary: Revoke all account sessions
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All refresh-token families revoked.
  /v1/me/sessions/{sessionID}:
    parameters:
      - $ref: "#/components/parameters/SessionID"
    delete:
      tags: [Authentication]
      operationId: revokeSession
      summary: Revoke one account session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked.
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/me/policy-acceptances:
    post:
      tags: [Policies]
      operationId: acceptPolicies
      summary: Accept current policy versions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy_acceptances]
              properties:
                policy_acceptances:
                  type: array
                  minItems: 1
                  items:
                    $ref: "#/components/schemas/PolicyAcceptance"
      responses:
        "204":
          description: Policies accepted.
        "401":
          $ref: "#/components/responses/Error"
        "428":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionID:
      name: sessionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    OpaqueToken:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [token]
            properties:
              token:
                type: string
                writeOnly: true
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    PolicyAcceptance:
      type: object
      required: [kind, version]
      properties:
        kind:
          type: string
        version:
          type: string
...
    Session:
      type: object
      required: [id, user_agent, created_at, expires_at]
      properties:
        id:
          type: string
          format: uuid
        user_agent:
          type: string
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: [string, "null"]
          format: date-time
        expires_at:
          type: string
          format: date-time
    SessionListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Session"
...